Smart Switch 6 & 7 Updates
As of the April 10, 2023 SmartThings update, devices with both switch and multi-level switch capabilities  will have the multi-level capability prioritized by the default driver to turn the device on or off. Since the multi-level capability only controls the indicator light, we introduce a custom handler to handle the on / off behavior.